Deploy to Pre-production

+++ packages/api/src/handlers/graphql/mutation/index.js
@@ -5,6 +5,7 @@ const {
   DeleteItemCommand,
   QueryCommand,
   BatchWriteItemCommand,
+  BatchGetItemCommand,
 } = require("@aws-sdk/client-dynamodb");
 const { marshall, unmarshall } = require("@aws-sdk/util-dynamodb");
 const { v4 } = require("uuid");
@@ -323,6 +324,88 @@ async function tagFileLabel(data) {
   return response;
 }
 
+async function bulkDeleteBackground(data) {
+  let response = {};
+  try {
+    let backgroundId = data.id;
+    const arrBackgroundItems = [];
+    const arrCompanyBackgroundItems = [];
+    const arrBackgroundIds = [];
+
+    for (var a = 0; a < backgroundId.length; a++) {
+      var bId = { id: backgroundId[a] };
+
+      arrBackgroundIds.push(bId);
+
+      arrBackgroundItems.push({
+        DeleteRequest: {
+          Key: marshall(bId),
+        },
+      });
+
+      const companyBackgroundParams = {
+        TableName: "ClientMatterBackgroundTable",
+        IndexName: "byBackground",
+        KeyConditionExpression: "backgroundId = :backgroundId",
+        ExpressionAttributeValues: marshall({
+          ":backgroundId": backgroundId[a],
+        }),
+      };
+
+      const companyBackgroundCommand = new QueryCommand(
+        companyBackgroundParams
+      );
+      const companyBackgroundResult = await client.send(
+        companyBackgroundCommand
+      );
+
+      for (var a = 0; a < companyBackgroundResult.Items.length; a++) {
+        var companyBackgroundId = { id: companyBackgroundResult.Items[a].id };
+        arrCompanyBackgroundItems.push({
+          DeleteRequest: {
+            Key: companyBackgroundId,
+          },
+        });
+      }
+    }
+
+    const backgroundParams = {
+      RequestItems: {
+        BackgroundsTable: arrBackgroundItems,
+      },
+    };
+
+    const backgroundCommand = new BatchWriteItemCommand(backgroundParams);
+    const backgroundResult = await client.send(backgroundCommand);
+
+    if (backgroundResult) {
+      const deleteCompanyBackgroundParams = {
+        RequestItems: {
+          ClientMatterBackgroundTable: arrCompanyBackgroundItems,
+        },
+      };
+
+      const deleteCompanyBackgroundCommand = new BatchWriteItemCommand(
+        deleteCompanyBackgroundParams
+      );
+      const deleteCompanyBackgroundResult = await client.send(
+        deleteCompanyBackgroundCommand
+      );
+
+      response = deleteCompanyBackgroundResult ? arrBackgroundIds : {};
+    }
+  } catch (e) {
+    response = {
+      error: e.message,
+      errorStack: e.stack,
+      statusCode: 500,
+    };
+    console.log(response);
+  }
+
+  return response;
+}
+
 async function updateLabel(id, data) {
   let response = {};
   try {
@@ -574,18 +657,74 @@ export async function deleteBackground(id) {
 
     const filterClientMatterBackgroundId = clientMatterBackgroundId[0];
 
-    const deleateClientMatterBackgroundCommand = new DeleteItemCommand({
+    const deleteClientMatterBackgroundCommand = new DeleteItemCommand({
       TableName: "ClientMatterBackgroundTable",
       Key: { id: filterClientMatterBackgroundId },
     });
 
-    const deleateClientMatterBackgroundResult = await client.send(
-      deleateClientMatterBackgroundCommand
+    const deleteClientMatterBackgroundResult = await client.send(
+      deleteClientMatterBackgroundCommand
     );
 
-    if (deleateClientMatterBackgroundResult) {
+    if (deleteClientMatterBackgroundResult) {
       const command = new DeleteItemCommand({
         TableName: "BackgroundsTable",
+        Key: marshall({ id }),
+      });
+      const request = await client.send(command);
+
+      response = request ? { id: id } : {};
+    }
+  } catch (e) {
+    response = {
+      error: e.message,
+      errorStack: e.stack,
+      statusCode: 500,
+    };
+    console.log(response);
+  }
+
+  return response;
+}
+
+export async function deleteClientMatter(id) {
+  let response = {};
+
+  try {
+    const companyClientMatterParams = {
+      TableName: "CompanyClientMatterTable",
+      IndexName: "byClientMatter",
+      KeyConditionExpression: "clientMatterId = :clientMatterId",
+      ExpressionAttributeValues: marshall({
+        ":clientMatterId": id,
+      }),
+    };
+
+    const companyClientMatterCommand = new QueryCommand(
+      companyClientMatterParams
+    );
+    const companyClientMatterResult = await client.send(
+      companyClientMatterCommand
+    );
+
+    const companyClientMatterId = companyClientMatterResult.Items.map(
+      (i) => i.id
+    );
+
+    const filterCompanyClientMatterId = companyClientMatterId[0];
+
+    const deleteCompanyClientMatterCommand = new DeleteItemCommand({
+      TableName: "CompanyClientMatterTable",
+      Key: { id: filterCompanyClientMatterId },
+    });
+
+    const deleteCompanyClientMatterResult = await client.send(
+      deleteCompanyClientMatterCommand
+    );
+
+    if (deleteCompanyClientMatterResult) {
+      const command = new DeleteItemCommand({
+        TableName: "ClientMatterTable",
         Key: marshall({ id }),
       });
       const request = await client.send(command);
@@ -690,6 +829,10 @@ const resolvers = {
     clientMatterCreate: async (ctx) => {
       return await createClientMatter(ctx.arguments);
     },
+    clientMatterDelete: async (ctx) => {
+      const { id } = ctx.arguments;
+      return await deleteClientMatter(id);
+    },
     backgroundCreate: async (ctx) => {
       return await createBackground(ctx.arguments);
     },
@@ -708,6 +851,9 @@ const resolvers = {
     backgroundDelete: async (ctx) => {
       const { id } = ctx.arguments;
       return await deleteBackground(id);
+    },
+    backgroundBulkDelete: async (ctx) => {
+      return await bulkDeleteBackground(ctx.arguments);
     },
   },
 };
